Add generated register API
As we implemented uart, we found some flaws with the existing scheme.
Specifically, writing individual fields of a packed struct (u32) on the
raspberry pi pico will result in incorrect modifications due to the
narrow write behavior (ctrl-f narrow in rp2040 datasheet)

Add a new API that fulfills the following constraints
* At runtime is equal cost to masking and setting values by hand
* Reserved fields are untouched by modification
* Hard to misuse
* Allows working with names/enums instead of manual bit shifting

This ended up being easiest to fulfil with a generated API. To avoid
changing registers, we need to read/modify/write. This modification
needs to be composable at comptime, and paramaterized. This means we
need functions that generate appropriate mask/value pairs + a way to
compose them. And the best way to make these functions is with codegen